Adding Role based access

+++ app/services/base_service.py
@@ -5,11 +5,13 @@ class BaseService:
     def __init__(self, model):
         self.model = model
 
-    def get_all_sorted(self):
-        return self.model.query.order_by(asc(self.model.id)).all()
+    def get_all_sorted(self, include_deleted=False):
+        query = self.model.query if include_deleted else self.model.get_active()
+        return query.order_by(asc(self.model.id)).all()
 
-    def get_by_id(self, id):
-        return self.model.query.get(id)
+    def get_by_id(self, id, include_deleted=False):
+        query = self.model.query if include_deleted else self.model.get_active()
+        return query.get(id)
 
     def create(self, **kwargs):
         instance = self.model(**kwargs)
@@ -25,9 +27,20 @@ class BaseService:
             db.session.commit()
         return instance
 
-    def delete(self, id):
+    def delete(self, id, soft=True):
         instance = self.get_by_id(id)
         if instance:
-            db.session.delete(instance)
+            if soft:
+                instance.soft_delete()
+                db.session.commit()
+            else:
+                db.session.delete(instance)
+                db.session.commit()
+        return instance
+
+    def restore(self, id):
+        instance = self.get_by_id(id, include_deleted=True)
+        if instance:
+            instance.restore()
             db.session.commit()
         return instance
